+++ libraries/dl-streamer/scripts/optimizer/README.md
@@ -6,20 +6,21 @@ Currently the DLSOptimizer focuses mainly on DL Streamers elements, specifically
 
 Multi-stream pipelines are also currently not supported.
 
-# Usage
+# Using the optimizer as a tool
 ```
-python3 optimizer.py [OPT] -- PIPELINE
+python3 . [OPT] -- PIPELINE
 
 Options:
     --search-duration SEARCH_DURATION   How long should the optimizer search for better pipelines
     --sample-duration SAMPLE_DURATION   How long should every pipeline be sampled for performance
+    --log-level LEVEL                   Configure the logging detail level
 ```
 
 Increasing the search duration will increase the chances of discovering more performant pipelines. Increasing the sample duration will improve the stability of the search, but less pipelines will potentially be explored. 
 
-# Example
+## Example
 ```
- python3 optimizer.py -- urisourcebin buffer-size=4096 uri=https://videos.pexels.com/video-files/1192116/1192116-sd_640_360_30fps.mp4 ! decodebin ! gvadetect model=/home/optimizer/models/public/yolo11s/INT8/yolo11s.xml ! queue ! gvawatermark ! vah264enc ! h264parse ! mp4mux ! fakesink
+ python3 . -- urisourcebin buffer-size=4096 uri=https://videos.pexels.com/video-files/1192116/1192116-sd_640_360_30fps.mp4 ! decodebin ! gvadetect model=/home/optimizer/models/public/yolo11s/INT8/yolo11s.xml ! queue ! gvawatermark ! vah264enc ! h264parse ! mp4mux ! fakesink
 [__main__] [    INFO] - GStreamer initialized successfully
 [__main__] [    INFO] - GStreamer version: 1.26.6
 [__main__] [    INFO] - Detected GPU Device
@@ -41,3 +42,34 @@ In this case the optimizer started with a pipeline that ran at ~45fps, and found
  - configuring the `gvadetect` element to use GPU for processing
  - setting the `batch-size` parameter to 2
  - setting the `nireq` parameter to 2
+
+# Using the optimizer as a library
+
+The easiest way of importing the optimizer into your scripts is to include it in your `PYTHONPATH` environment variable:
+```export PYTHONPATH=/opt/intel/dlstreamer/scripts/optimizer```
+
+Targets which are exported in order to facilitate usage inside of scripts:
+
+ - `preprocess_pipeline(pipeline) -> processed_pipeline` - Perform quick search and replace for known combinations of elements with more performant alternatives.
+   - `pipeline`: String - A string containing a valid DL Streamer pipeline.
+   - `processed_pipeline`: String - A string containing the pipelines with all relevant substitutions.
+ - `get_optimized_pipeline(pipeline, search_duration, sample_duration) -> (optimized_pipeline, fps)` - Perform the optimization process on a pipeline.
+   - `pipeline`: String - A string containing a valid DL Streamer pipeline.
+   - `search_duration`: Int - The duration of searching for optimized pipelines in seconds, default `300`.
+   - `sample_duration`: Int - The duration of sampling each candidate pipeline in seconds, default `10`.
+   - `optimized_pipeline`: String - The best performing pipeline that has been found during the search.
+   - `fps`: Float - The measured fps of the best perfmorming pipeline.
+
+## Example
+```python
+from optimizer import get_optimized_pipeline
+
+pipeline = "urisourcebin buffer-size=4096 uri=https://videos.pexels.com/video-files/1192116/1192116-sd_640_360_30fps.mp4 ! decodebin ! gvadetect model=/home/optimizer/models/public/yolo11s/INT8/yolo11s.xml ! queue ! gvawatermark ! vah264enc ! h264parse ! mp4mux ! fakesink"
+
+optimized_pipeline, fps = get_optimized_pipeline(pipeline)
+print("Best discovered pipeline: " + optimized_pipeline)
+print("Measured fps: " + fps)
+```
+
+# Controling the measurement
+The point at which performance is being measured can be controlled by pre-emptively inserting a `gvafpscounter` element into your pipeline definition. For pipelines which lack such an element, the measurement is done after the last inference element supported by the optimizer tool.
