+++ libraries/dl-streamer/scripts/optimizer/__init__.py
@@ -1,0 +1,7 @@
+# ==============================================================================
+# Copyright (C) 2025-2025 Intel Corporation
+#
+# SPDX-License-Identifier: MIT
+# ==============================================================================
+
+__all__ = [get_optimized_pipeline, preprocess_pipeline]
