+++ libraries/dl-streamer/scripts/optimizer/optimizer.py
@@ -3,14 +3,13 @@
 #
 # SPDX-License-Identifier: MIT
 # ==============================================================================
+from preprocess import preprocess_pipeline
+from processors.inference import add_device_suggestions, add_batch_suggestions, add_nireq_suggestions, parse_element_parameters # pylint: disable=line-too-long
 
-import argparse
 import time
 import logging
 import itertools
 import os
-import re
-import subprocess
 
 import gi
 gi.require_version("Gst", "1.0")
@@ -19,7 +18,6 @@ from gi.repository import Gst
 ####################################### Init ######################################################
 
 Gst.init()
-logging.basicConfig(level=logging.INFO, format="[%(name)s] [%(levelname)8s] - %(message)s")
 logger = logging.getLogger(__name__)
 logger.info("GStreamer initialized successfully")
 gst_version = Gst.version()
@@ -28,88 +26,53 @@ logger.info("GStreamer version: %d.%d.%d",
             gst_version.minor,
             gst_version.micro)
 
-####################################### Utils #####################################################
+####################################### Main Logic ################################################
 
-def parse_element_parameters(element):
-    parameters = element.strip().split(" ")
-    del parameters[0]
-    parsed_parameters = {}
-    for parameter in parameters:
-        parts = parameter.split("=")
-        parsed_parameters[parts[0]] = parts[1]
+# Steps of pipeline optimization:
+# 1. Measure the baseline pipeline's performace.
+# 2. Pre-process the pipeline to cover cases where we're certain of the best alternative.
+# 3. Prepare a set of processors providing alternatives for elements.
+# 3. Run the processors in sequence and test their effect on the pipeline.
+# 5. For every processor create a cartesian product of the suggestions
+#    and start running the combinations to measure performance.
+# 6. Any time a better pipeline is found, save it and its performance information.
+# 7. Return the best discovered pipeline.
+def get_optimized_pipeline(pipeline, search_duration = 300, sample_duration = 10):
+    pipeline = pipeline.split("!")
 
-    return parsed_parameters
-
-def assemble_parameters(parameters):
-    result = ""
-    for parameter, value in parameters.items():
-        result = result + parameter + "=" + value + " "
-
-    return result
-
-def log_parameters_of_interest(pipeline):
-    for element in pipeline:
-        if "gvadetect" in element:
-            parameters = parse_element_parameters(element)
-            logger.info("Found Gvadetect, device: %s, batch size: %s, nireqs: %s",
-                        parameters.get("device", "not set"),
-                        parameters.get("batch-size", "not set"),
-                        parameters.get("nireq", "not set"))
-
-        if "gvaclassify" in element:
-            parameters = parse_element_parameters(element)
-            logger.info("Found Gvaclassify, device: %s, batch size: %s, nireqs: %s",
-                        parameters.get("device", "not set"),
-                        parameters.get("batch-size", "not set"),
-                        parameters.get("nireq", "not set"))
-
-###################################### System Scanning ############################################
-
-def scan_system():
-    context = {"GPU": False,
-               "NPU": False}
-
-    # check for presence of GPU
+    # Measure the performance of the original pipeline
     try:
-        gpu_query = subprocess.run(["dpkg", "-l", "intel-opencl-icd"],
-                                   stderr=subprocess.DEVNULL,
-                                   stdout=subprocess.DEVNULL,
-                                   check=False)
-        gpu_dir = os.listdir("/dev/dri")
-        for file in gpu_dir:
-            if "render" in file and gpu_query.returncode == 0:
-                context["GPU"] = True
+        fps = sample_pipeline(pipeline, sample_duration)
+    except Exception as e:
+        logger.error("Pipeline failed to start, unable to measure fps: %s", e)
+        raise RuntimeError("Provided pipeline is not valid") from e
 
-    # can happen on missing directory, signifies no GPU support
-    except Exception: # pylint: disable=broad-exception-caught
-        pass
+    logger.info("FPS: %f.2", fps)
 
-    if context["GPU"]:
-        logger.info("Detected GPU Device")
-    else:
-        logger.info("No GPU Device detected")
+    # Make pipeline definition portable across inference devices.
+    # Replace elements with known better alternatives.
+    pipeline = "!".join(pipeline)
+    pipeline = preprocess_pipeline(pipeline)
+    pipeline = pipeline.split("!")
 
-    # check for presence of NPU
-    try:
-        npu_query = subprocess.run(["dpkg", "-l", "intel-driver-compiler-npu"],
-                                   stderr=subprocess.DEVNULL,
-                                   stdout=subprocess.DEVNULL,
-                                   check=False)
-        npu_dir = os.listdir("/dev/accel/")
-        for file in npu_dir:
-            if "accel" in file and npu_query.returncode == 0:
-                context["NPU"] = True
+    processors = [
+        add_device_suggestions,
+        add_batch_suggestions,
+        add_nireq_suggestions,
+    ]
 
-    # can happen on missing directory, signifies no NPU support
-    except Exception: # pylint: disable=broad-exception-caught
-        pass
+    search_end_time = time.time() + search_duration
+    for processor in processors:
+        remaining_duration = search_end_time - time.time()
+        if search_end_time <= time.time():
+            break
 
-    if context["NPU"]:
-        logger.info("Detected NPU Device")
-    else:
-        logger.info("No NPU Device detected")
+        suggestions = prepare_suggestions(pipeline)
+        processor(suggestions)
+        pipeline, fps = explore_pipelines(suggestions, fps, remaining_duration, sample_duration)
 
-    return context
+    # Reconstruct the pipeline as a single string and return it.
+    return "!".join(pipeline), fps
 
 ##################################### Pipeline Running ############################################
 
@@ -121,6 +84,8 @@ def explore_pipelines(suggestions, base_fps, search_duration, sample_duration):
     best_fps = base_fps
     for combination in combinations:
         combination = list(combination)
+        # re-slice the pipeline to handle cases where a suggestion added multiple elements
+        combination = "!".join(combination).split("!")
         log_parameters_of_interest(combination)
 
         try:
@@ -142,15 +107,17 @@ def explore_pipelines(suggestions, base_fps, search_duration, sample_duration):
 def sample_pipeline(pipeline, sample_duration):
     pipeline = pipeline.copy()
 
-    # check if there is an fps counter after the last inference element
-    for i, element in enumerate(reversed(pipeline)):
-        # exit early if one is found before other elements
+    # check if there is an fps counter in the pipeline, add one otherwise
+    has_fps_counter = False
+    for element in pipeline:
         if "gvafpscounter" in element:
-            break
+            has_fps_counter = True
 
-        # add one if no counter was found before inference elements
-        if "gvadetect" in element or "gvaclassify" in element:
-            pipeline.insert(len(pipeline) - i, "gvafpscounter")
+    if not has_fps_counter:
+        for i, element in enumerate(reversed(pipeline)):
+            if "gvadetect" in element or "gvaclassify" in element:
+                pipeline.insert(len(pipeline) - i, " gvafpscounter " )
+                break
 
     pipeline = "!".join(pipeline)
     logger.debug("Testing: %s", pipeline)
@@ -162,7 +129,10 @@ def sample_pipeline(pipeline, sample_duration):
 
     bus = pipeline.get_bus()
 
-    pipeline.set_state(Gst.State.PLAYING)
+    ret = pipeline.set_state(Gst.State.PLAYING)
+    _, state, _ = pipeline.get_state(Gst.CLOCK_TIME_NONE)
+    logger.debug("Pipeline state: %s, %s", state, ret)
+
     terminate = False
     start_time = time.time()
     while not terminate:
@@ -173,6 +143,7 @@ def sample_pipeline(pipeline, sample_duration):
         _, state, _ = pipeline.get_state(Gst.CLOCK_TIME_NONE)
         if state == Gst.State.READY:
             pipeline.set_state(Gst.State.NULL)
+            process_bus(bus)
             del pipeline
             raise RuntimeError("Pipeline not healthy, terminating early")
 
@@ -180,8 +151,18 @@ def sample_pipeline(pipeline, sample_duration):
         if cur_time - start_time > sample_duration:
             terminate = True
 
-    pipeline.set_state(Gst.State.NULL)
+    ret = pipeline.set_state(Gst.State.NULL)
+    logger.debug("Setting pipeline to NULL: %s", ret)
+    _, state, _ = pipeline.get_state(Gst.CLOCK_TIME_NONE)
+    logger.debug("Pipeline state: %s", str(state))
+    process_bus(bus)
 
+    del pipeline
+    fps = fps_counter.get_property("avg-fps")
+    logger.debug("Sampled fps: %f.2", fps)
+    return fps
+
+def process_bus(bus):
     # Process any messages from the bus
     message = bus.pop()
     while message is not None:
@@ -198,121 +179,7 @@ def sample_pipeline(pipeline, sample_duration):
             logger.error("Other message: %s", str(message))
         message = bus.pop()
 
-    del pipeline
-    fps = fps_counter.get_property("avg-fps")
-    logger.debug("Sampled fps: %f.2", fps)
-    return fps
-
-######################################## Preprocess ###############################################
-
-preprocessing_rules = {
-    "vaapi": "va",
-    "memory:VASurface": "memory:VAMemory",
-    r"! capsfilter caps=video/x-raw\(memory:VAMemory\) !(.*(gvadetect|gvaclassify))": r"!\1",
-    r"! video/x-raw\(memory:VAMemory\) !(.*(gvadetect|gvaclassify))": r"!\1",
-    r"parsebin ! vah\w*dec": "decodebin3",
-    r"\w*parse ! vah\w*dec": "decodebin3",
-    r"\bdecodebin\b": "decodebin3",
-}
-
-def preprocess_pipeline(pipeline):
-    pipeline = "!".join(pipeline)
-    for pattern, replacement in preprocessing_rules.items():
-        if re.search(pattern, pipeline):
-            pipeline = re.sub(pattern, replacement, pipeline)
-    pipeline = pipeline.split("!")
-    return pipeline
-
-#################################### Gvadetect & Gvaclassify ######################################
-
-def add_device_suggestions(suggestions, context):
-    for suggestion in suggestions:
-        for element in ["gvadetect", "gvaclassify"]:
-            if element in suggestion[0]:
-                parameters = parse_element_parameters(suggestion[0])
-
-                if context["GPU"] & (parameters.get("device", "") != "GPU"):
-                    parameters["device"] = "GPU"
-                    parameters["pre-process-backend"] = "va-surface-sharing"
-                    suggestion.append(f" {element} {assemble_parameters(parameters)}")
-
-                if context["NPU"] & (parameters.get("device", "") != "NPU"):
-                    parameters["device"] = "NPU"
-                    parameters["pre-process-backend"] = "va"
-                    suggestion.append(f" {element} {assemble_parameters(parameters)}")
-
-                parameters["device"] = "CPU"
-                parameters["pre-process-backend"] = "opencv"
-                suggestion.append(f" {element} {assemble_parameters(parameters)}")
-
-def add_batch_suggestions(suggestions, _):
-    batches = [1, 2, 4, 8, 16, 32]
-    for suggestion in suggestions:
-        for element in ["gvadetect", "gvaclassify"]:
-            if element in suggestion[0]:
-                parameters = parse_element_parameters(suggestion[0])
-                for batch in batches:
-                    parameters["batch-size"] = str(batch)
-                    suggestion.append(f" {element} {assemble_parameters(parameters)}")
-
-
-def add_nireq_suggestions(suggestions, _):
-    nireqs = range(1, 9)
-    for suggestion in suggestions:
-        for element in ["gvadetect", "gvaclassify"]:
-            if element in suggestion[0]:
-                parameters = parse_element_parameters(suggestion[0])
-                for nireq in nireqs:
-                    parameters["nireq"] = str(nireq)
-                    suggestion.append(f" {element} {assemble_parameters(parameters)}")
-
-####################################### Main Logic ################################################
-
-# Steps of pipeline optimization:
-# 1. Measure the baseline pipeline's performace.
-# 2. Pre-process the pipeline to cover cases where we're certain of the best alternative.
-# 3. Prepare a set of processors providing alternatives for elements.
-# 3. Run the processors in sequence and test their effect on the pipeline.
-# 5. For every processor create a cartesian product of the suggestions
-#    and start running the combinations to measure performance.
-# 6. Any time a better pipeline is found, save it and its performance information.
-# 7. Return the best discovered pipeline.
-def get_optimized_pipeline(pipeline, search_duration = 300, sample_duration = 10):
-    context = scan_system()
-
-    pipeline = pipeline.split("!")
-
-    # Measure the performance of the original pipeline
-    try:
-        fps = sample_pipeline(pipeline, sample_duration)
-    except Exception as e:
-        logger.error("Pipeline failed to start, unable to measure fps: %s", e)
-        raise RuntimeError("Provided pipeline is not valid") from e
-
-    logger.info("FPS: %f.2", fps)
-
-    # Make pipeline definition portable across inference devices.
-    # Replace elements with known better alternatives.
-    pipeline = preprocess_pipeline(pipeline)
-
-    processors = [
-        add_device_suggestions,
-        add_batch_suggestions,
-        add_nireq_suggestions,
-    ]
-
-    search_end_time = time.time() + search_duration
-    for processor in processors:
-        remaining_duration = search_end_time - time.time()
-        if search_end_time <= time.time():
-            break
-
-        suggestions = prepare_suggestions(pipeline)
-        processor(suggestions, context)
-        pipeline, fps = explore_pipelines(suggestions, fps, remaining_duration, sample_duration)
-
-    # Reconstruct the pipeline as a single string and return it.
-    return "!".join(pipeline), fps
+####################################### Utils #####################################################
 
 def prepare_suggestions(pipeline):
     # Prepare the suggestions structure
@@ -328,29 +195,18 @@ def prepare_suggestions(pipeline):
         suggestions.append([element])
     return suggestions
 
+def log_parameters_of_interest(pipeline):
+    for element in pipeline:
+        if "gvadetect" in element:
+            parameters = parse_element_parameters(element)
+            logger.info("Found Gvadetect, device: %s, batch size: %s, nireqs: %s",
+                        parameters.get("device", "not set"),
+                        parameters.get("batch-size", "not set"),
+                        parameters.get("nireq", "not set"))
 
-def main():
-    parser = argparse.ArgumentParser(
-        prog="DLStreamer Pipeline Optimization Tool",
-        description="Use this tool to try and find versions of your pipeline that will run with increased performance." # pylint: disable=line-too-long
-    )
-    parser.add_argument("--search-duration", default=300,
-                        help="Duration of time which should be spent searching for optimized pipelines (default: %(default)ss)") # pylint: disable=line-too-long
-    parser.add_argument("--sample-duration", default=10,
-                        help="Duration of sampling individual pipelines. Longer duration should offer more stable results (default: %(default)ss)") # pylint: disable=line-too-long
-    parser.add_argument("pipeline", nargs="+",
-                        help="Pipeline to be analyzed")
-    args=parser.parse_args()
-
-    pipeline = " ".join(args.pipeline)
-
-    try:
-        best_pipeline, best_fps = get_optimized_pipeline(pipeline,
-                                                         args.search_duration,
-                                                         args.sample_duration)
-        logger.info("\nBest found pipeline: %s \nwith fps: %f.2", best_pipeline, best_fps)
-    except Exception as e: # pylint: disable=broad-exception-caught
-        logger.error("Failed to optimize pipeline: %s", e)
-
-if __name__ == "__main__":
-    main()
+        if "gvaclassify" in element:
+            parameters = parse_element_parameters(element)
+            logger.info("Found Gvaclassify, device: %s, batch size: %s, nireqs: %s",
+                        parameters.get("device", "not set"),
+                        parameters.get("batch-size", "not set"),
+                        parameters.get("nireq", "not set"))
